+++ src/Passes/Source/Rules/Notation/BasicBlock.cpp
@@ -1,30 +1,26 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-#include "Rules/Notation/Notation.hpp"
-#include "Rules/Operands/Instruction.hpp"
-
 #include "Llvm/Llvm.hpp"
+#include "Rules/Notation/Notation.hpp"
+#include "Rules/Patterns/Instruction.hpp"
 
 #include <unordered_map>
 #include <vector>
 
-namespace microsoft
+namespace microsoft {
+namespace quantum {
+namespace notation {
+
+using IOperandPrototypePtr = std::shared_ptr<IOperandPrototype>;
+
+IOperandPrototypePtr basicBlock()
 {
-namespace quantum
-{
-    namespace notation
-    {
+  auto ret = std::make_shared<BasicBlockPattern>();
 
-        using IOperandPrototypePtr = std::shared_ptr<IOperandPrototype>;
+  return static_cast<IOperandPrototypePtr>(ret);
+}
 
-        IOperandPrototypePtr basicBlock()
-        {
-            auto ret = std::make_shared<BasicBlockPattern>();
-
-            return static_cast<IOperandPrototypePtr>(ret);
-        }
-
-    } // namespace notation
-} // namespace quantum
-} // namespace microsoft
+}  // namespace notation
+}  // namespace quantum
+}  // namespace microsoft
