+++ src/Passes/Source/Rules/Notation/Store.cpp
@@ -1,32 +1,28 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-#include "Rules/Notation/Notation.hpp"
-#include "Rules/Operands/Instruction.hpp"
-
 #include "Llvm/Llvm.hpp"
+#include "Rules/Notation/Notation.hpp"
+#include "Rules/Patterns/Instruction.hpp"
 
 #include <unordered_map>
 #include <vector>
 
-namespace microsoft
+namespace microsoft {
+namespace quantum {
+namespace notation {
+
+using IOperandPrototypePtr = std::shared_ptr<IOperandPrototype>;
+
+IOperandPrototypePtr store(IOperandPrototypePtr const &target, IOperandPrototypePtr const &value)
 {
-namespace quantum
-{
-    namespace notation
-    {
+  auto ret = std::make_shared<StorePattern>();
 
-        using IOperandPrototypePtr = std::shared_ptr<IOperandPrototype>;
+  ret->addChild(target);
+  ret->addChild(value);
+  return static_cast<IOperandPrototypePtr>(ret);
+}
 
-        IOperandPrototypePtr store(IOperandPrototypePtr const& target, IOperandPrototypePtr const& value)
-        {
-            auto ret = std::make_shared<StorePattern>();
-
-            ret->addChild(target);
-            ret->addChild(value);
-            return static_cast<IOperandPrototypePtr>(ret);
-        }
-
-    } // namespace notation
-} // namespace quantum
-} // namespace microsoft
+}  // namespace notation
+}  // namespace quantum
+}  // namespace microsoft
