+++ src/Passes/Source/Rules/Patterns/AnyPattern.hpp
